separate ha-tests from charm-tests
